Fix packaging
