--- src/main/java/command/ExitCommand.java
@@ -1,16 +1,0 @@
-package command;
-
-public class ExitCommand extends Command {
-    public ExitCommand() {
-
-    }
-
-    @Override
-    public String toString() {
-        return "bye";
-    }
-
-    public boolean isExit() {
-        return true;
-    }
-}
