--- src/main/java/exception/UnknownCommandException.java
@@ -1,17 +1,0 @@
-package exception;
-
-public class UnknownCommandException extends SpheneException {
-    public UnknownCommandException(String command) {
-        super(command, null);
-    }
-
-    @Override
-    public String toString() {
-        return "Unknown command error from command: " + this.getCommand();
-    }
-
-    @Override
-    public String getMessage() {
-        return "I don't know how to complete the request '" + this.getCommand() + "'.";
-    }
-}
