+++ src/main/java/sphene/command/ExitCommand.java
@@ -1,0 +1,16 @@
+package sphene.command;
+
+public class ExitCommand extends Command {
+    public ExitCommand() {
+
+    }
+
+    @Override
+    public String toString() {
+        return "bye";
+    }
+
+    public boolean isExit() {
+        return true;
+    }
+}
