+++ src/main/java/sphene/command/MarkCommand.java
@@ -1,0 +1,26 @@
+package sphene.command;
+
+import sphene.component.Storage;
+import sphene.component.TaskList;
+import sphene.component.Ui;
+import sphene.exception.OutOfListRangeException;
+import sphene.exception.SaveException;
+
+public class MarkCommand extends Command {
+    private final int index;
+
+    public MarkCommand(int index) {
+        this.index = index;
+    }
+
+    @Override
+    public String toString() {
+        return "mark " + this.index;
+    }
+
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) throws SaveException, OutOfListRangeException {
+        tasks.markTask(index);
+        storage.store(tasks.serialize());
+    }
+}
