+++ src/main/java/sphene/exception/EmptyFieldException.java
@@ -1,0 +1,19 @@
+package sphene.exception;
+
+public class EmptyFieldException extends SpheneException {
+    private final String field;
+    public EmptyFieldException(String command, String params, String field) {
+        super(command, params);
+        this.field = field;
+    }
+
+    @Override
+    public String toString() {
+        return "Field '" + this.field + "' in sphene.command '" + this.getCommand() + "' should not be empty";
+    }
+
+    @Override
+    public String getMessage() {
+        return "The field '" + this.field + "' in the request '" + this.getCommand() + "' should not be empty.";
+    }
+}
