+++ src/main/java/sphene/task/ToDo.java
@@ -1,0 +1,17 @@
+package sphene.task;
+
+public class ToDo extends Task {
+    public ToDo(String content) {
+        super(content);
+    }
+
+    @Override
+    public String toString() {
+        return "[T]" + super.toString();
+    }
+
+    @Override
+    public String serialize() {
+        return "T" + "," + super.serialize();
+    }
+}
