--- src/main/java/task/ToDo.java
@@ -1,17 +1,0 @@
-package task;
-
-public class ToDo extends Task {
-    public ToDo(String content) {
-        super(content);
-    }
-
-    @Override
-    public String toString() {
-        return "[T]" + super.toString();
-    }
-
-    @Override
-    public String serialize() {
-        return "T" + "," + super.serialize();
-    }
-}
